Asynchronous poll deletion
Poll deletion can take quite some time. Especially when deleting all
polls of a user (which can be up to 200 at a time).

This lead to super long-running requests for those deletions, which in
turn lead to race-conditions with other requests.

For instance, users could still operate on polls that were going to be
deleted in a few minutes. A few exceptions popped up because of this
exact problem.

To fix this, a new flag has been introduced on the poll: `delete`.
Polls with this flag can no longer be interacted with.
It's basically a mutex that prevents users to interact with it until
it's finally deleted.

The whole process of deleting polls and their messages has been moved
to a background task.