Allow users to set a duration time in minutes for a poll due date
Adds a menu settings option to set a specific due time duration in
minutes from current datetime. For example, if a user set 25 minutes
of due time duration and the current time is 15/10/2022 10:10 HS
it will set the due datetime on 15/10/2022 10:35 HS.
The duration time is added on the due datetime data, so any
edge case of a date is cover by the addition.